pyproject.toml
[tool.poetry]
name = "spacefy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
django-widget-tweaks = "^1.5.0"
pillow = "^10.1.0"
django-allauth = "^0.58.2"
django-debug-toolbar = "^4.2.0"
redis = "^5.0.1"
celery = "^5.3.6"
pytest-factoryboy = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
